+++ packages/lib/server.ts
@@ -6,3 +6,5 @@ export * from './src/dictionary';
 export * from './src/database';
 export * from './src/registry';
 export * from './src/themeLayout';
+export * from './src/dispatch';
+export * from './src/emails/BasicTemplate';