+++ packages/providers/index.ts
@@ -9,6 +9,7 @@ export * from './src/ToastProvider';
 export * from './src/SupportProvider';
 export * from './src/EnvironmentProvider';
 export * from './src/ShortcutsProvider';
+export * from './src/DispatchProvider';
 export * from './src/UserProvider';
 export * from './src/DictionaryProvider';
 export * from './src/NotificationProvider';