+++ packages/ui/index.ts
@@ -10,3 +10,4 @@ export * from './src/Modal';
 export * from './src/PreventLossConfirmation';
 export * from './src/CollapsibleCard';
 export * from './src/Input';
+export * from './src/MultiSelector';