+++ apps/app/CHANGELOG.md
@@ -1,5 +1,15 @@
 # app
 
+## 2.1.2
+
+### Patch Changes
+
+- Updated dependencies
+    - @xernerx/ui@0.2.7
+    - @xernerx/providers@0.3.4
+    - @xernerx/components@0.2.11
+    - @xernerx/feedback@0.0.12
+
 ## 2.1.1
 
 ### Patch Changes