+++ apps/auth/CHANGELOG.md
@@ -1,5 +1,15 @@
 # auth
 
+## 0.6.3
+
+### Patch Changes
+
+- Updated dependencies
+    - @xernerx/ui@0.2.7
+    - @xernerx/providers@0.3.4
+    - @xernerx/components@0.2.11
+    - @xernerx/feedback@0.0.12
+
 ## 0.6.2
 
 ### Patch Changes