+++ apps/www/CHANGELOG.md
@@ -1,5 +1,15 @@
 # www
 
+## 0.5.4
+
+### Patch Changes
+
+- Updated dependencies
+    - @xernerx/ui@0.2.7
+    - @xernerx/providers@0.3.4
+    - @xernerx/components@0.2.11
+    - @xernerx/feedback@0.0.12
+
 ## 0.5.3
 
 ### Patch Changes